+++ src/components/Tile/index.tsx
@@ -1,0 +1,286 @@
+import React from 'react';
+import classNames from 'classnames';
+
+import Button from '../Button';
+import OverlayWrapper from '../OverlayWrapper';
+import Popover from '../Popover';
+import Dropdown from '../Dropdown';
+import ButtonIcon from '../ButtonIcon';
+import Icon from 'react-oui-icons';
+
+import { greyDark } from '../../tokens/forimport/index.es';
+
+
+const renderDropdownActivator = (
+  { onClick, buttonRef } // eslint-disable-line
+) => (
+  <ButtonIcon
+    style="plain"
+    iconFill="default"
+    iconName="ellipsis"
+    buttonRef={buttonRef}
+    onClick={onClick}
+  />
+);
+
+const renderTileActions = (
+  items: {
+    text: string;
+    description?: string;
+    onClick: (...args: any[]) => any;
+  }[],
+  testSection?: string
+) => {
+  return (
+    <Dropdown
+      renderActivator={renderDropdownActivator}
+      placement={'bottom-end'}
+      key="dropdown"
+    >
+      <Dropdown.Contents direction={'right'}>
+        {items.map((item, idx) => (
+          <Dropdown.ListItem key={`dropdown-item-${idx}`}>
+            <Dropdown.BlockLink
+              onClick={item.onClick}
+              testSection={
+                testSection && `${testSection}-dropdown-block-link-${idx}`
+              }
+            >
+              <Dropdown.BlockLinkText text={item.text} />
+              {item.description && (
+                <Dropdown.BlockLinkSecondaryText
+                  secondaryText={item.description}
+                />
+              )}
+            </Dropdown.BlockLink>
+          </Dropdown.ListItem>
+        ))}
+      </Dropdown.Contents>
+    </Dropdown>
+  );
+};
+
+export type TileProps = {
+  /**
+   * Description of the item for this reference Tile
+   */
+  description?: string;
+
+  /**
+   * Optional properties to include when using Tile
+   * in DragAndDrop component
+   */
+  dragHandleProps?: object;
+
+  /**
+   * Optional dropdown items to add to right side of Tile
+   * Descriptions are optional, title and onClick required
+   */
+  dropdownItems?: {
+    text: string;
+    description?: string;
+    onClick: (...args: any[]) => any;
+  }[];
+
+  /**
+   * Whether or not this Tile has margin on the ends
+   * True by default
+   */
+  hasSpacing?: boolean;
+
+  /**
+   * Whether or not this Tile has a warning popover
+   */
+  hasWarning?: boolean;
+
+  /**
+   * Whether or not this Tile includes a drag handle
+   */
+  isDraggable?: boolean;
+
+  /**
+   * Whether or not this Tile is currently selected/active
+   */
+  isSelected?: boolean;
+
+  /**
+   * Function to call when copy button is clicked
+   * Supplying this function adds a copy button
+   * to the tile
+   */
+  onCopy?: (...args: any[]) => any;
+
+  /**
+   * Function to call when delete button is clicked
+   * Supplying this function adds a delete button
+   * to the tile
+   */
+  onDismiss?: (...args: any[]) => any;
+
+  /**
+   * Function to call when edit button is clicked
+   * Supplying this function adds an edit button
+   * to the tile
+   */
+  onEdit?: (...args: any[]) => any;
+
+  /**
+   * Function to call when the main area of the Tile is clicked
+   * If function is not supplied, main content of the Tile
+   * will not be clickable (div instead of a button)
+   */
+  onTileClick?: (...args: any[]) => any;
+
+  /**
+   * Optional number used to indicate the order of Tiles
+   */
+  order?: number;
+
+  /**
+   * Optional string used to indicate status before action items
+   */
+  status?: string;
+
+  /**
+   * String to use for testing
+   */
+  testSection?: string;
+
+  /**
+   * Title of the Tile, required
+   */
+  name: string;
+
+  /**
+   * Whether or not the title of this Tile displays in monospace font
+   */
+  usesMonospaceName?: boolean;
+
+  /**
+   * The content of the warning popover
+   */
+  warningContent?: any;
+
+  /**
+   * The title of the warning popover
+   */
+  warningTitle?: string;
+};
+
+const Tile = ({
+  description,
+  dragHandleProps,
+  dropdownItems,
+  hasSpacing = true,
+  hasWarning = false,
+  isDraggable = false,
+  isSelected = false,
+  name,
+  onCopy,
+  onDismiss,
+  onEdit,
+  onTileClick,
+  order,
+  status,
+  testSection,
+  usesMonospaceName = false,
+  warningContent = '',
+  warningTitle,
+}: TileProps) => {
+  const tileContent = (
+    <>
+      <p
+        className={classNames('text--left flush', {
+          monospace: usesMonospaceName,
+        })}
+      >
+        {name}
+      </p>
+      <p className="text--left muted flush micro wrap-text">{description}</p>
+    </>
+  );
+  return (
+    <div
+      className={classNames('oui-tile flex flex-align--center soft', {
+        'oui-tile--selected': isSelected,
+        'push-half--ends': hasSpacing,
+      })}
+      data-test-section={testSection}
+    >
+      {order && (
+        <span className="oui-tile__order-number milli text--right push--right">
+          {order}
+        </span>
+      )}
+      {isDraggable && (
+        <div className="oui-tile__drag-handle push--right" {...dragHandleProps}>
+          <Icon name="hamburger" fill={greyDark} />
+        </div>
+      )}
+      {hasWarning && (
+        <div className="push--right push-half--top display--inline-block">
+          <OverlayWrapper
+            behavior="hover"
+            horizontalAttachment="center"
+            overlay={
+              <Popover
+                title={warningTitle}
+                testSection={`${testSection}-warning-popover`}
+              >
+                {warningContent}
+              </Popover>
+            }
+            shouldHideOnClick={true}
+            verticalAttachment="bottom"
+          >
+            <Icon name="exclamation" />
+          </OverlayWrapper>
+        </div>
+      )}
+      {onTileClick ? (
+        <Button style="unstyled" width="full" onClick={onTileClick}>
+          {tileContent}
+        </Button>
+      ) : (
+        <div className="width--1-1 line--1">{tileContent}</div>
+      )}
+      <div className="flex flex-justified--between flex-align--center push--left">
+        {status && (
+          <p className="muted micro flush--bottom weight--bold push-half--sides">
+            {status}
+          </p>
+        )}
+        {onCopy && (
+          <ButtonIcon
+            iconName="clipboard"
+            size="medium"
+            style="plain"
+            iconFill="default"
+            onClick={onCopy}
+          />
+        )}
+        {onEdit && (
+          <ButtonIcon
+            iconName="pencil"
+            size="medium"
+            style="plain"
+            iconFill="default"
+            onClick={onEdit}
+          />
+        )}
+        {onDismiss && (
+          <ButtonIcon
+            iconName="trash"
+            size="medium"
+            style="plain"
+            iconFill="default"
+            onClick={onDismiss}
+          />
+        )}
+        {dropdownItems && renderTileActions(dropdownItems, testSection)}
+      </div>
+    </div>
+  );
+};
+
+export default Tile;